% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgp_utilities.r
\name{analyze_legs}
\alias{analyze_legs}
\title{Analyze Legs}
\usage{
analyze_legs(pgp)
}
\arguments{
\item{pgp}{A pgp object}
}
\value{
list A list with a detail of each leg including the start and end
  dates, the number of periods, the type (Risky or Safe) and the gross and
  net returns.  The Summary is a data frame with rows for the long term gains
  and losses, short-term gains and losses (for the Risky asset) and a row for
  the safe investments.  The summary for each row includes the number of
  legs, how many periods in total were invested in those legs and the
  cumulative gross and net returns.
}
\description{
Useful for analyzing taxes and the degree of turnover.  A leg is a period
over which a position in either the risky or safe asset is held, that is from
the time it is bought until it is sold.
}
\examples{
analyze_legs(pgp)
}

