% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgp_utilities.r
\name{avg_px}
\alias{avg_px}
\title{Average Price - adds a moving average of the prices}
\usage{
avg_px(pgp)
}
\arguments{
\item{pgp}{A PGP class object with at least a data frame with the RiskyPx column}
}
\value{
pgp object
}
\description{
Average Price - adds a moving average of the prices
}
\examples{
avg_px(pgp)
}

