% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.r
\name{load_dfa_csv}
\alias{load_dfa_csv}
\title{Load data from DFA CSV file}
\usage{
load_dfa_csv(fn)
}
\arguments{
\item{fn}{Filename including path of csv file}
}
\value{
list Risk is name of risky asset, Safe is name of safe.  data is
  dataframe containing RiskyRet, SafeRet, RiskyPx, and SafePx which are
  returns and prices.
}
\description{
File should have Date in m/d/Y, followed by a risky and safe asset return in
decimal format example "1/1/1926",-0.0385,0.00338
}
\examples{
load_dfa_csv("periodic returns")
}

