% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgp_utilities.r
\name{period_returns}
\alias{period_returns}
\title{Period Returns - generate periodic returns (monthly, daily depending on frequency of data)}
\usage{
period_returns(pgp)
}
\arguments{
\item{pgp}{pgp object}
}
\value{
pgp object
}
\description{
Calculates both gross and net returns using the commission rate.
}
\examples{
period_returns(pgp)
}

