% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgp_utilities.r
\name{pgp_backtest}
\alias{pgp_backtest}
\title{pgp_backtest}
\usage{
pgp_backtest(pgp)
}
\arguments{
\item{pgp}{A pgp object}
}
\value{
pgp updated pgp with results
}
\description{
Runs a backtest on the data.
}
\examples{
pgp_backtest(pgp)
}

