% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgp_utilities.r
\name{set_params}
\alias{set_params}
\title{Set Parameters of PGP object}
\usage{
set_params(pgp, width, commission)
}
\arguments{
\item{pgp}{}

\item{width}{Number of prices in moving average}

\item{commission}{Commission rate in percent.}
}
\value{
pgp object
}
\description{
Set Parameters of PGP object
}
\examples{
set_params(pgp,11,0.1)
}

